Back-End Finalizado

+++ src/users/users.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Users } from './user.entity';
@@ -13,8 +13,15 @@ export class UsersService {
 
   async findAllUsers(): Promise<Users[]> {
     const users = await this.usersRepository.find();
-
     return users;
+  }
+
+  async findUserId(id: string): Promise<Users> {
+    const user = await this.usersRepository.findOneBy({ id });
+    if (!user) {
+      throw new NotFoundException(`User with ID ${id} not found`);
+    }
+    return user;
   }
 
   async createUsers(user: UserDomain): Promise<UserDomain> {
@@ -31,11 +38,9 @@ export class UsersService {
     return await this.usersRepository.save(updateUser);
   }
 
-  async deleteUser(id: string): Promise<any>{
-    const deleteUser = await this.usersRepository.delete(id);
-
-    if(!deleteUser) throw new Error ("Usuário Não Deletado");
-
-    return deleteUser
+  async deleteUser(id: string): Promise<void> {
+    const user = await this.usersRepository.findOneBy({ id });
+    if (!user) throw new NotFoundException('Usuário não encontrado');
+    await this.usersRepository.delete(id);
   }
 }
